+++ packages/core/src/graph/graph.ts
@@ -817,6 +817,18 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
   }
 
   /**
+   * Focus on the passed items
+   * @param {Item[]} items Items you want to focus on
+   * @param {boolean} zoomToFit Wether to zoom on the passed items
+   * @param {boolean} animate Wether to animate the transition
+   * @param {GraphAnimateConfig} animateCfg Animation configuration
+   */
+  public focusItems(items: Item[], zoomToFit?: boolean, animate?: boolean, animateCfg?: GraphAnimateConfig): void {
+    const viewController: ViewController = this.get('viewController');
+    viewController.focusItems(items, zoomToFit, animate, animateCfg);
+  }
+
+  /**
    * 自动重绘
    * @internal 仅供内部更新机制调用，外部根据需求调用 render 或 paint 接口
    */
@@ -2509,7 +2521,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
    * 若 cfg 含有 type 字段或为 String 类型，且与现有布局方法不同，则更换布局
    * 若 cfg 不包括 type ，则保持原有布局方法，仅更新布局配置项
    */
-  public updateLayout(cfg: any, align?: 'center' | 'begin', alignPoint?: IPoint, stack: boolean = true): void {
+  public updateLayout(cfg: any = {}, align?: 'center' | 'begin', alignPoint?: IPoint, stack: boolean = true): void {
     const layoutController = this.get('layoutController');
 
     if (isString(cfg)) {
@@ -2528,7 +2540,7 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
       // translate to point coordinate system
       toPoint = this.getPointByCanvas(toPoint.x, toPoint.y);
 
-      const forceTypes = ['force', 'gForce', 'fruchterman'];
+      const forceTypes = ['force', 'gForce', 'fruchterman', 'force2'];
 
       // if it is force layout, only center takes effect, and assign center force
       if (forceTypes.includes(cfg.type) || (!cfg.type && forceTypes.includes(layoutController?.layoutType))) {
@@ -2554,10 +2566,11 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
     const oriLayoutCfg = { ...this.get('layout') };
     const layoutCfg: any = {};
     Object.assign(layoutCfg, oriLayoutCfg, cfg);
+    if (cfg.pipes && !cfg.type) delete layoutCfg.type;
+    else if (!cfg.pipes && layoutCfg.type) delete layoutCfg.pipes;
     this.set('layout', layoutCfg);
 
     if (!layoutController) return;
-
     if (
       layoutController.isLayoutTypeSame(layoutCfg) &&
       layoutCfg.gpuEnabled === oriLayoutCfg.gpuEnabled
@@ -2812,8 +2825,8 @@ export default abstract class AbstractGraph extends EventEmitter implements IAbs
         let selfEndModel = selfEnd.getModel();
         // find the nearest visible ancestor
         while (!selfEnd.isVisible()) {
-          const { parentId: selfEndPId, comboId: selfEndCId } = otherEndModel;
-          const selfEndParentId = selfEndPId || selfEndCId;
+          const { parentId: selfEndPId, comboId: selfEndCId } = selfEndModel;
+          const selfEndParentId = (selfEndPId || selfEndCId) as string;
           selfEnd = this.findById(selfEndParentId) as ICombo;
           if (!selfEnd || !selfEndParentId) {
             return; // if all the ancestors of the oppsite are all hidden, ignore the edge